refactor: move plugin logic from command to plugin
