+++ src/MultiTekla.Plugins/ModelCommand.cs
@@ -1,0 +1,10 @@
+using CliFx;
+
+namespace MultiTekla.Plugins;
+
+[Command("model", Description = "Model related commands")]
+public class ModelCommand : ICommand
+{
+    public ValueTask ExecuteAsync(IConsole console)
+        => default;
+}
